chore(jans-cedarling): refactor raw config
